+++ apps/commerce-api/src/test/java/com/loopers/application/like/ProductLikeFacadeIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.loopers.application;
+package com.loopers.application.like;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
