+++ apps/commerce-api/src/test/java/com/loopers/domain/member/EmailTest.java
@@ -1,11 +1,9 @@
-package com.loopers.domain;
+package com.loopers.domain.member;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import com.loopers.domain.member.Email;
 
 class EmailTest {
     @DisplayName("Email 객체를 테스트한다.")
